# docs/swagger.yaml
basePath: /
definitions:
  models.Account:
    properties:
      balance:
        type: integer
      currency:
        type: integer
      id:
        type: string
      isCredit:
        type: boolean
      userId:
        type: string
    type: object
  models.Customer:
    properties:
      blocked:
        type: boolean
      fullname:
        type: string
      id:
        type: string
      login:
        type: string
      password:
        type: string
      timecreated:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  license:
    name: Apache 2.0
  title: PetBank API
  version: "1.0"
paths:
  /balance:
    post:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
      summary: create
      tags:
      - Accounts
  /blockaccount:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: userid
        in: path
        name: userid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
      summary: create
      tags:
      - Accounts
  /getaccountdetails:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: userid
        in: path
        name: userid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
      summary: create
      tags:
      - Accounts
  /getallaccounts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Account'
            type: array
      summary: create
      tags:
      - Accounts
  /getallusers:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Customer'
            type: array
      summary: get all
      tags:
      - Users
  /newaccount:
    post:
      consumes:
      - application/json
      parameters:
      - description: Userid, IsCredit, Balance, Currency
        in: body
        name: data
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
      summary: create
      tags:
      - Accounts
  /signin:
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Customer'
      summary: Sign in
      tags:
      - Users
  /signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Customer'
      summary: create
      tags:
      - Users
  /user/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Customer'
      summary: Get user details with id
      tags:
      - Users
  /user/{id}/block:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Block user by himself with id
      tags:
      - Users
swagger: "2.0"